CH-444/CH-581 deleted packed 'impl' jira[why doesn't work @Validated] https://cozyhometeam.atlassian.net/browse/CH-581?atlOrigin=eyJpIjoiNmI1YjgyMWY0MTQ2NDVlZmJjMGM0ZDNiYzIxZmEyOGEiLCJwIjoiaiJ9
