+++ src/main/java/com/cozyhome/onlineshop/productservice/controller/ColorControllerImpl.java
@@ -1,4 +1,4 @@
-package com.cozyhome.onlineshop.productservice.controller.impl;
+package com.cozyhome.onlineshop.productservice.controller;
 
 import com.cozyhome.onlineshop.dto.ColorDto;
 import com.cozyhome.onlineshop.productservice.controller.swagger.CommonApiResponses;
