+++ src/main/java/com/cozyhome/onlineshop/productservice/controller/ProductControllerImpl.java
@@ -1,4 +1,4 @@
-package com.cozyhome.onlineshop.productservice.controller.impl;
+package com.cozyhome.onlineshop.productservice.controller;
 
 import com.cozyhome.onlineshop.dto.ProductDto;
 import com.cozyhome.onlineshop.dto.ProductStatusDto;
